Separate client and client configuration structs into their own modules
In preparation for the introduction of features in the next commit.
